switched id's to 64 bit

+++ Core/Include/NoobWarrior/EmuDb/EmuDb.h
@@ -167,41 +167,41 @@ public:
     /* Generic item functions */
     SqlDb::Response AddBlob(const std::vector<unsigned char> &data, std::string *hashOutput = nullptr);
     SqlDb::Response AddItem(ItemType type, SqlRow row);
-    SqlDb::Response UpdateItem(ItemType, int id, SqlRow row);
-    SqlDb::Response DeleteItem(ItemType, int id);
+    SqlDb::Response UpdateItem(ItemType, int64_t id, SqlRow row);
+    SqlDb::Response DeleteItem(ItemType, int64_t id);
 
     /* Asset functions */
-    SqlDb::Response AttachDataToAsset(int id, int version, const std::vector<unsigned char> &data);
-    SqlDb::Response DetachDataFromAsset(int id, int version);
+    SqlDb::Response AttachDataToAsset(int64_t id, int version, const std::vector<unsigned char> &data);
+    SqlDb::Response DetachDataFromAsset(int64_t id, int version);
 
-    SqlDb::Response AttachBlobHashToAsset(int id, int version, const std::string &hash);
-    SqlDb::Response DetachBlobHashFromAsset(int id, int version, const std::string &hash);
+    SqlDb::Response AttachBlobHashToAsset(int64_t id, int version, const std::string &hash);
+    SqlDb::Response DetachBlobHashFromAsset(int64_t id, int version, const std::string &hash);
 
-    SqlDb::Response AttachHistoricalDataToAsset(int id, SqlRow row);
-    SqlDb::Response DetachHistoricalDataFromAsset(int id, SqlRow row);
+    SqlDb::Response AttachHistoricalDataToAsset(int64_t id, SqlRow row);
+    SqlDb::Response DetachHistoricalDataFromAsset(int64_t id, SqlRow row);
 
-    SqlDb::Response AttachMicrotransactionDataToAsset(int id, SqlRow row);
-    SqlDb::Response DetachMicrotransactionDataFromAsset(int id, SqlRow row);
+    SqlDb::Response AttachMicrotransactionDataToAsset(int64_t id, SqlRow row);
+    SqlDb::Response DetachMicrotransactionDataFromAsset(int64_t id, SqlRow row);
 
-    SqlDb::Response AddThumbnailToPlace(int id, int imageId);
-    SqlDb::Response RemoveThumbnailFromPlace(int id, int imageId);
+    SqlDb::Response AddThumbnailToPlace(int64_t id, int64_t imageId);
+    SqlDb::Response RemoveThumbnailFromPlace(int64_t id, int64_t imageId);
 
-    SqlDb::Response RenderThumbnailForAsset(int id, int version = 0);
+    SqlDb::Response RenderThumbnailForAsset(int64_t id, int version = 0);
 
     /* Bundle functions */
-    SqlDb::Response AddAssetToBundle(int bundleId, int assetId);
-    SqlDb::Response RemoveAssetFromBundle(int bundleId, int assetId);
+    SqlDb::Response AddAssetToBundle(int64_t bundleId, int64_t assetId);
+    SqlDb::Response RemoveAssetFromBundle(int64_t bundleId, int64_t assetId);
 
-    SqlDb::Response AddAssetToOutfit(int outfitId, int assetId);
-    SqlDb::Response RemoveAssetFromOutfit(int outfitId, int assetId);
+    SqlDb::Response AddAssetToOutfit(int64_t outfitId, int64_t assetId);
+    SqlDb::Response RemoveAssetFromOutfit(int64_t outfitId, int64_t assetId);
 
     /* User functions */
-    SqlDb::Response AddAssetToUserCharacter(int userId, int assetId);
-    SqlDb::Response RemoveAssetFromUserCharacter(int userId, int assetId);
+    SqlDb::Response AddAssetToUserCharacter(int64_t userId, int64_t assetId);
+    SqlDb::Response RemoveAssetFromUserCharacter(int64_t userId, int64_t assetId);
 
     AssetRepository* GetAssetRepository();
 
-    std::vector<unsigned char> RetrieveImageData(const std::string &tableName, int id);
+    std::vector<unsigned char> RetrieveImageData(const std::string &tableName, int64_t id);
 
     template<typename T>
     static T GetValueFromColumnIndex(sqlite3_stmt *stmt, int columnIndex) {
